+++ 2ndTermProject/EarthArmy/EarthGunnery.h
@@ -6,7 +6,7 @@
 class EarthGunnery : public ArmyUnit
 {
 public:
-	EarthGunnery(Game* p, int HP, int pow, int ID_, int cap, int Tj, Unit U = EG);
+	EarthGunnery(Game* p, double HP, double pow, int ID_, int cap, int Tj, Unit U = EG);
 	virtual bool Attack();
 	friend ostream& operator<<(ostream& COUT, EarthGunnery* Passed_AU);
 };
