+++ 2ndTermProject/EarthArmy/EarthSoldier.cpp
@@ -1,7 +1,7 @@
 #include "EarthSoldier.h"
 #include "../Game.h"
 
-EarthSoldier::EarthSoldier(Game* p, int HP, int pow, int ID_, int cap, int _Tj, Unit U)
+EarthSoldier::EarthSoldier(Game* p, double HP, double pow, int ID_, int cap, int _Tj, Unit U)
 	:ArmyUnit(p, HP, pow, ID_, cap, _Tj, U)
 {
 }
@@ -20,7 +20,7 @@ bool EarthSoldier::Attack() //PHASE2
 				if (ES->Get_Ta() == -1)
 					ES->Set_Ta(pGame->Get_Tj());
 
-				int dmg = int((float)GetPower() * (GetHealth() / 100.0) / (float)sqrt(ES->GetHealth()));
+				double dmg = GetPower() * (GetHealth() / 100.0) / sqrt(ES->GetHealth());
 				ES->SetHealth(ES->GetHealth() - dmg);
 				templist.enqueue(ES);
 			}
@@ -59,7 +59,7 @@ bool EarthSoldier::Attack() //PHASE2
 				if (AS->Get_Ta() == -1)
 					AS->Set_Ta(pGame->Get_Tj());
 
-				int dmg = int((float)GetPower() * (GetHealth() / 100.0) / (float)sqrt(AS->GetHealth()));
+				double dmg = GetPower() * (GetHealth() / 100.0) / sqrt(AS->GetHealth());
 				AS->SetHealth(AS->GetHealth() - dmg);
 				templist.enqueue(AS);
 			}
